Move deletedelay to core
