+++ redbot/cogs/mod/mod.py
@@ -11,7 +11,6 @@ from redbot.core.i18n import Translator, cog_i18n
 from .casetypes import CASETYPES
 from .events import Events
 from .kickban import KickBanMixin
-from .movetocore import MoveToCore
 from .mutes import MuteMixin
 from .names import ModInfo
 from .slowmode import Slowmode
@@ -19,7 +18,7 @@ from .settings import ModSettings
 
 _ = T_ = Translator("Mod", __file__)
 
-__version__ = "1.1.0"
+__version__ = "1.2.0"
 
 
 class CompositeMetaClass(type(commands.Cog), type(ABC)):
@@ -36,7 +35,6 @@ class Mod(
     ModSettings,
     Events,
     KickBanMixin,
-    MoveToCore,
     MuteMixin,
     ModInfo,
     Slowmode,
@@ -113,6 +111,16 @@ class Mod(
             ).format(prefix=prefix)
             self.bot.loop.create_task(self.bot.send_to_owners(msg))
             await self.settings.version.set(__version__)
+        if await self.settings.version() < "1.2.0":
+            prefixes = await self.bot.get_valid_prefixes()
+            prefix = re.sub(rf"<@!?{self.bot.user.id}>", f"@{self.bot.user.name}", prefixes[0])
+            msg = _(
+                "Delete delay settings have been moved. "
+                "Please use `{prefix}movedeletedelay` if "
+                "you were previously using these functions."
+            ).format(prefix=prefix)
+            self.bot.loop.create_task(self.bot.send_to_owners(msg))
+            await self.settings.version.set(__version__)
 
     @commands.command()
     @commands.is_owner()
@@ -127,3 +135,15 @@ class Mod(
             await self.bot._config.channel_from_id(channel_id).ignored.set(settings["ignored"])
             await self.settings.channel_from_id(channel_id).clear()
         await ctx.send(_("Ignored channels and guilds restored."))
+
+    @commands.command()
+    @commands.is_owner()
+    async def movedeletedelay(self, ctx: commands.Context) -> None:
+        """
+            Move deletedelay settings to core
+        """
+        all_guilds = await self.settings.all_guilds()
+        for guild_id, settings in all_guilds.items():
+            await self.bot._config.guild_from_id(guild_id).delete_delay.set(settings["delete_delay"])
+            await self.settings.guild_from_id(guild_id).delete_delay.clear()
+        await ctx.send(_("Delete delay settings restored."))
